requirements.txt
azure-functions
azure-eventhub
azure-storage-blob==12.19.1
azure-core==1.29.4
cryptography==43.0.3
